refactor: Change `get_size` return type to `u64`

+++ wnfs-wasm/src/fs/private/file.rs
@@ -102,7 +102,7 @@ impl PrivateFile {
                 .await
                 .map_err(error("Cannot determine file size"))?;
 
-            Ok(value!(size))
+            Ok(value!(size as usize))
         }))
     }
